* Updated property definitions, so that the app builds in 32-bit mode

+++ UI/CMMsxDisplayView.m
@@ -207,8 +207,8 @@ static CVReturn renderCallback(CVDisplayLinkRef displayLink,
             {
                 emulator.scanlines = scanlineAmount;
 #ifdef DEBUG
-                NSLog(@"Resetting scanlines to %ld%% - screen not large enough",
-                      scanlineAmount);
+                NSLog(@"Resetting scanlines to %d%% - screen not large enough",
+                      (int)scanlineAmount);
 #endif
             }
         }
